Libify pipecolor
